@@ -1,74 +1,76 @@
 {
     "name": "@anglr/rest",
-    "version": "4.0.14",
+    "version": "5.0.0-beta.1",
     "description": "Angular module representing rest services",
     "main": "dist/index.js",
     "typings": "dist/index.d.ts",
     "scripts": {
         "build": "npm run ngc",
         "postbuild": "copyfiles -u 2 dist/src/**/*.* dist && copyfiles -u 2 dist/testing/**/*.* testing && rimraf dist/src dist/dist dist/node_modules dist/testing",
-        "clean": "rimraf dist testing tests",
+        "clean": "rimraf dist testing tests docs",
         "clean:deps": "rimraf package-lock.json node_modules",
         "clean:all": "npm run clean && npm run clean:deps",
         "ngc": "ngc",
         "prepublishOnly": "npm run bundle",
-        "postpublish": "npm run clean",
+        "postpublish": "npm run clean && npm run docs",
         "republish": "npm publish --force",
-        "bundle": "npm run clean && npm run build",
+        "bundle": "npm run clean && npm run build && copyfiles -u 1 src/**/*.html dist && nnrh *.js -d dist",
         "deploy:stable": "npp -b",
         "deploy:next": "npp -p -b -s \"beta\" -t \"next\"",
         "deploy:stable:local": "npp -b --registry http://local:4873",
         "deploy:next:local": "npp -p -b -s \"beta\" -t \"next\" --registry http://local:4873",
+        "docs": "compodoc src -d docs -n '@anglr/rest' -p tsconfig.json --disablePrivate --disableInternal --disableLifeCycleHooks",
+        "postdocs": "copyfiles -u 1 docs/**/*.* \\\\DOCS-SERVER\\docs\\ng-rest && npm run clean",
         "pretest": "webpack",
         "test": "karma start karma.conf.js"
     },
     "author": "kukjevov",
     "license": "MIT",
     "peerDependencies": {
-        "@angular/core": ">=5.0.0",
-        "@angular/common": ">=5.0.0",
-        "@angular/platform-browser": "^5.0.0",
-        "rxjs": ">=5.5.0",
-        "@anglr/common": ">=4.0.0",
-        "jquery-param": "^0.2.0",
+        "@angular/core": ">=6.0.0",
+        "@angular/common": ">=6.0.0",
+        "rxjs": ">=6.0.0",
+        "@anglr/common": ">=5.0.0-beta.1",
+        "jquery-param": "^1.0.1",
         "crypto-js": "^3.1.9-1"
     },
     "devDependencies": {
-        "@angular/core": "^5.0.0",
-        "@angular/common": "^5.0.0",
-        "@angular/router": "^5.0.0",
-        "@angular/forms": "^5.0.0",
-        "rxjs": "^5.5.0",
-        "@anglr/common": "^4.0.0",
-        "@types/jasmine": "~2.6.0",
-        "@types/crypto-js": "^3.1.37",
-        "@types/node": "^8.0.41",
-        "@anglr/types": "^1.1.8",
-        "@angular/compiler-cli": "^5.0.0",
-        "@angular/compiler": "^5.0.0",
-        "@angular/language-service": "^5.0.0",
-        "@angular/platform-browser-dynamic": "^5.0.0",
-        "@angular/platform-browser": "^5.0.0",
-        "phantomjs-prebuilt": "~2.1.15",
+        "@angular/core": "^6.0.0",
+        "@angular/common": "^6.0.0",
+        "rxjs": "^6.0.0",
+        "@anglr/common": "^5.0.0-beta.1",
+        "@anglr/types": "1.1.12",
+        "@types/jasmine": "~2.8.7",
+        "@types/crypto-js": "^3.1.39",
+        "@angular/forms": "^6.0.0",
+        "@angular/router": "^6.0.0",
+        "@angular/compiler-cli": "^6.0.0",
+        "@angular/compiler": "^6.0.0",
+        "@angular/language-service": "^6.0.0",
+        "@angular/platform-browser-dynamic": "^6.0.0",
+        "@angular/platform-browser": "^6.0.0",
+        "phantomjs-prebuilt": "~2.1.16",
         "karma-cli": "~1.0.1",
-        "karma": "~1.7.1",
-        "karma-jasmine": "~1.1.0",
+        "karma": "~2.0.2",
+        "karma-jasmine": "~1.1.2",
         "karma-phantomjs-launcher": "~1.0.4",
         "karma-chrome-launcher": "~2.2.0",
-        "karma-webpack": "~2.0.5",
-        "karma-jasmine-html-reporter": "~0.2.2",
-        "jasmine": "~2.8.0",
-        "awesome-typescript-loader": "~3.4.1",
+        "karma-webpack": "~3.0.0",
+        "karma-jasmine-html-reporter": "~1.1.0",
+        "jasmine": "~3.1.0",
+        "awesome-typescript-loader": "~5.0.0",
         "angular2-template-loader": "~0.6.2",
         "raw-loader": "~0.5.1",
-        "sass-loader": "~6.0.6",
-        "css-loader": "~0.28.7",
-        "file-loader": "~1.1.5",
-        "style-loader": "~0.19.0",
+        "sass-loader": "~7.0.1",
+        "css-loader": "~0.28.11",
+        "file-loader": "~1.1.11",
+        "style-loader": "~0.21.0",
+        "@compodoc/compodoc": "~1.0.9",
         "rimraf": "~2.6.2",
-        "copyfiles": "~1.2.0",
-        "typescript": "2.6.2",
-        "webpack": "3.8.1",
-        "npm-publish-package": "~1.2.0"
+        "copyfiles": "~2.0.0",
+        "typescript": "~2.7.2",
+        "webpack": "~4.8.1",
+        "npm-publish-package": "~1.2.0",
+        "npm-ng-require-html": "~1.0.0"
     }
 }
