@@ -1,8 +1,6 @@
 import 'core-js/es6';
 import 'core-js/es7/reflect';
-import 'rxjs/Observable';
-import 'rxjs/Subject';
-import 'rxjs/Subscription';
+import 'rxjs';
 
 import 'zone.js/dist/zone';
 import 'zone.js/dist/long-stack-trace-zone';
